Split bigquery
