+++ frictionless/plugins/bigquery/storage.py
@@ -5,166 +5,15 @@ import json
 import time
 from slugify import slugify
 from functools import partial
-from ..exception import FrictionlessException
-from ..resource import Resource
-from ..package import Package
-from ..storage import Storage
-from ..metadata import Metadata
-from ..dialect import Dialect
-from ..parser import Parser
-from ..plugin import Plugin
-from ..schema import Schema
-from ..field import Field
-from .. import helpers
-from .. import errors
-
-
-# NOTE:
-# We need to ensure that the way we detect bigquery service is good enough.
-# We don't want to be importing google and checking the type withouth a good reason
-
-
-# Plugin
-
-
-class BigqueryPlugin(Plugin):
-    """Plugin for BigQuery
-
-    API      | Usage
-    -------- | --------
-    Public   | `from frictionless.plugins.bigquery import BigqueryPlugin`
-    """
-
-    code = "bigquery"
-    status = "experimental"
-
-    def create_file(self, file):
-        if not file.scheme and not file.format and file.memory:
-            if helpers.is_type(file.data, "Resource"):
-                file.scheme = ""
-                file.format = "bigquery"
-                return file
-
-    def create_dialect(self, resource, *, descriptor):
-        if resource.format == "bigquery":
-            return BigqueryDialect(descriptor)
-
-    def create_parser(self, resource):
-        if resource.format == "bigquery":
-            return BigqueryParser(resource)
-
-    def create_storage(self, name, source, **options):
-        if name == "bigquery":
-            return BigqueryStorage(source, **options)
-
-
-# Dialect
-
-
-class BigqueryDialect(Dialect):
-    """Bigquery dialect representation
-
-    API      | Usage
-    -------- | --------
-    Public   | `from frictionless.plugins.bigquery import BigqueryDialect`
-
-    Parameters:
-        descriptor? (str|dict): descriptor
-        project (str): project
-        dataset? (str): dataset
-        table? (str): table
-
-    Raises:
-        FrictionlessException: raise any error that occurs during the process
-    """
-
-    def __init__(
-        self,
-        descriptor=None,
-        *,
-        project=None,
-        dataset=None,
-        prefix=None,
-        table=None,
-    ):
-        self.setinitial("project", project)
-        self.setinitial("dataset", dataset)
-        self.setinitial("prefix", prefix)
-        self.setinitial("table", table)
-        super().__init__(descriptor)
-
-    @Metadata.property
-    def project(self):
-        return self.get("project")
-
-    @Metadata.property
-    def dataset(self):
-        return self.get("dataset")
-
-    @Metadata.property
-    def prefix(self):
-        return self.get("prefix") or ""
-
-    @Metadata.property
-    def table(self):
-        return self.get("table")
-
-    # Metadata
-
-    metadata_profile = {  # type: ignore
-        "type": "object",
-        "required": ["table"],
-        "additionalProperties": False,
-        "properties": {
-            "project": {"type": "string"},
-            "dataset": {"type": "string"},
-            "prefix": {"type": "string"},
-            "table": {"type": "string"},
-        },
-    }
-
-
-# Parser
-
-
-class BigqueryParser(Parser):
-    """Bigquery parser implementation.
-
-    API      | Usage
-    -------- | --------
-    Public   | `from frictionless.plugins.bigquery import BigqueryParser`
-    """
-
-    supported_types = [
-        "string",
-        "number",
-    ]
-
-    # Read
-
-    def read_list_stream_create(self):
-        dialect = self.resource.dialect
-        storage = BigqueryStorage(self.resource.data, dialect=dialect)
-        resource = storage.read_resource(dialect.table)
-        self.resource.schema = resource.schema
-        with resource:
-            yield from resource.list_stream
-
-    # Write
-
-    # NOTE: this approach is questionable
-    def write_row_stream(self, resource):
-        source = resource
-        target = self.resource
-        storage = BigqueryStorage(self.resource.data, dialect=target.dialect)
-        if not target.dialect.table:
-            note = 'Please provide "dialect.table" for writing'
-            raise FrictionlessException(errors.StorageError(note=note))
-        source.name = target.dialect.table
-        storage.write_resource(source, force=True)
-
-
-# Storage
+from ...exception import FrictionlessException
+from ...resource import Resource
+from ...package import Package
+from ...storage import Storage
+from ...schema import Schema
+from ...field import Field
+from ... import helpers
+from ... import errors
+from .dialect import BigqueryDialect
 
 
 class BigqueryStorage(Storage):
